+++ BEPUphysics/BroadPhaseEntries/MobileCollidables/MobileCollidable.cs
@@ -23,7 +23,7 @@
         /// Updates the bounding box of the mobile collidable.
         ///</summary>
         ///<param name="dt">Timestep with which to update the bounding box.</param>
-        public abstract void UpdateBoundingBox(float dt);
+        public abstract void UpdateBoundingBox(double dt);
 
 
     }
